tests/wpt/tests/css/css-fonts/parsing/font-size-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-size with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-size-prop">
<meta name="assert" content="font-size supports only the grammar '<absolute-size> | <relative-size> | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-size', 'auto');
test_invalid_value('font-size', 'medium small');

test_invalid_value('font-size', '-10px');
test_invalid_value('font-size', '-20%');
</script>
</body>
</html>
